Added webpack
Added configuration for Webpack, which will allow for server side rendering with call to import statements.
*** npx webpack
This should be run anytime you are making a change to expose-components.js and adding new component files. They ensure thath the files in "~/dist" remain up to date.
Each time a new 'npx webpack' is run if a "~/dist/' file is updated a new instance of the file with a new hash is created. This has been kept in incase of errors of not having permission to overwrite files. Check npx log to ensure which is the latest version. Or delete all and re-run 'npx webpack'.
** make sure to run npm install and npm run build if creating a new repository or having node issues.

**The ReactJs.Net tutorial has added some opinionated libraries such as router, helmet, styledcomponents. I have left these in but as we near production builds will remove as needed.